made the image loading component separate for reuse
